Move EVC and CEV to core utils
