+++ src/promptflow/promptflow/_sdk/_submitter/utils.py
@@ -40,15 +40,12 @@ from promptflow._sdk._constants import (
 )
 from promptflow._sdk._errors import InvalidFlowError, RunOperationError
 from promptflow._sdk._load_functions import load_flow
-from promptflow._sdk._utils import (
-    _merge_local_code_and_additional_includes,
-    get_used_connection_names_from_dict,
-    update_dict_value_with_connections,
-)
+from promptflow._sdk._utils import _merge_local_code_and_additional_includes
 from promptflow._sdk.entities._flow import FlexFlow, Flow
 from promptflow._utils.flow_utils import dump_flow_dag, load_flow_dag
 from promptflow._utils.logger_utils import FileHandler, get_cli_sdk_logger
 from promptflow.contracts.flow import Flow as ExecutableFlow
+from promptflow.core._utils import get_used_connection_names_from_dict, update_dict_value_with_connections
 from promptflow.exceptions import UserErrorException
 
 logger = get_cli_sdk_logger()
