+++ src/promptflow/promptflow/core/_serving/flow_invoker.py
@@ -14,6 +14,11 @@ from promptflow.core._connection import _Connection
 from promptflow.core._serving._errors import UnexpectedConnectionProviderReturn, UnsupportedConnectionProvider
 from promptflow.core._serving.flow_result import FlowResult
 from promptflow.core._serving.utils import validate_request_data
+from promptflow.core._utils import (
+    override_connection_config_with_environment_variable,
+    resolve_connections_environment_variable_reference,
+    update_environment_variables_with_connections,
+)
 from promptflow.executor import FlowExecutor
 from promptflow.storage._run_storage import DefaultRunStorage
 
@@ -112,13 +117,6 @@ class FlowInvoker:
             self.connections.update(connections)
         else:
             raise UnsupportedConnectionProvider(connection_provider)
-
-        # TODO (3027983): remove connection related dependency from promptflow-core
-        from promptflow._sdk._utils import (
-            override_connection_config_with_environment_variable,
-            resolve_connections_environment_variable_reference,
-            update_environment_variables_with_connections,
-        )
 
         override_connection_config_with_environment_variable(self.connections)
         resolve_connections_environment_variable_reference(self.connections)
