+++ src/promptflow/promptflow/executor/flow_executor.py
@@ -426,7 +426,7 @@ class FlowExecutor:
         :return: A dictionary containing updated environment variables.
         :rtype: dict
         """
-        from promptflow._sdk._utils import update_environment_variables_with_connections
+        from promptflow.core._utils import update_environment_variables_with_connections
 
         return update_environment_variables_with_connections(connections)
